Moved part into footer. Also experieced weird issue with random files showing up in my visual studio?...
