+++ next-app/src/components/footer.js
@@ -1,0 +1,71 @@
+import React from "react";
+import styles from "@/styles/Home.module.css";
+import { Inter } from "next/font/google";
+
+const inter = Inter({ subsets: ["latin"] });
+
+
+const footer = () => {
+  return (
+    <div className={styles.grid}>
+      <a
+        href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+        className={styles.card}
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        <h2 className={inter.className}>
+          Docs <span>-&gt;</span>
+        </h2>
+        <p className={inter.className}>
+          Find in-depth information about Next.js features and&nbsp;API.
+        </p>
+      </a>
+
+      <a
+        href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+        className={styles.card}
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        <h2 className={inter.className}>
+          Learn <span>-&gt;</span>
+        </h2>
+        <p className={inter.className}>
+          Learn about Next.js in an interactive course with&nbsp;quizzes!
+        </p>
+      </a>
+
+      <a
+        href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+        className={styles.card}
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        <h2 className={inter.className}>
+          Templates <span>-&gt;</span>
+        </h2>
+        <p className={inter.className}>
+          Discover and deploy boilerplate example Next.js&nbsp;projects.
+        </p>
+      </a>
+
+      <a
+        href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+        className={styles.card}
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        <h2 className={inter.className}>
+          Deploy <span>-&gt;</span>
+        </h2>
+        <p className={inter.className}>
+          Instantly deploy your Next.js site to a shareable URL
+          with&nbsp;Vercel.
+        </p>
+      </a>
+    </div>
+  );
+};
+
+export default footer;
