+++ next-app/src/pages/index.js
@@ -1,6 +1,6 @@
 import Home from '../components/home'
 
-export default function HomePage() {  
+export default function HomePage() {
   return (
     <>
       <Home />
